Add WhitelistRequestAccessDialog as separate component
